Move KeyDimQuery to separate module
